+++ services/invitations/src/simcore_service_invitations/api/_invitations.py
@@ -1,16 +1,19 @@
 import logging
 from datetime import datetime
-from typing import Any, Optional
+from typing import Any
 
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import HTTPBasicCredentials
-from pydantic import BaseModel, Field, HttpUrl
+from models_library.api_schemas_invitations.invitations import (
+    ApiEncryptedInvitation,
+    ApiInvitationContent,
+    ApiInvitationContentAndLink,
+    ApiInvitationInputs,
+)
 
 from ..core.settings import ApplicationSettings
 from ..invitations import (
     InvalidInvitationCode,
-    InvitationContent,
-    InvitationInputs,
     create_invitation_link,
     extract_invitation_code_from,
     extract_invitation_content,
@@ -33,38 +36,6 @@ _INPUTS_EXAMPLE: dict[str, Any] = {
 }
 
 
-class _ApiInvitationInputs(InvitationInputs):
-    class Config:
-        schema_extra = {"example": _INPUTS_EXAMPLE}
-
-
-class _ApiInvitationContent(InvitationContent):
-    class Config:
-        schema_extra = {
-            "example": {
-                **_INPUTS_EXAMPLE,
-                "created": "2023-01-11 13:11:47.293595",
-            }
-        }
-
-
-class _InvitationContentAndLink(_ApiInvitationContent):
-    invitation_url: HttpUrl = Field(..., description="Invitation link")
-
-    class Config:
-        schema_extra = {
-            "example": {
-                **_INPUTS_EXAMPLE,
-                "created": "2023-01-11 12:11:47.293595",
-                "invitation_url": "https://foo.com/#/registration?invitation=1234",
-            }
-        }
-
-
-class _EncryptedInvitation(BaseModel):
-    invitation_url: HttpUrl = Field(..., description="Invitation link")
-
-
 #
 # ROUTE HANDLERS
 #
@@ -73,13 +44,13 @@ router = APIRouter()
 
 @router.post(
     "/invitations",
-    response_model=_InvitationContentAndLink,
+    response_model=ApiInvitationContentAndLink,
     response_model_by_alias=False,
 )
 async def create_invitation(
-    invitation_inputs: _ApiInvitationInputs,
+    invitation_inputs: ApiInvitationInputs,
     settings: ApplicationSettings = Depends(get_settings),
-    _credentials: Optional[HTTPBasicCredentials] = Depends(get_validated_credentials),
+    _credentials: HTTPBasicCredentials | None = Depends(get_validated_credentials),
 ):
     """Generates a new invitation code and returns its content and an invitation link"""
 
@@ -88,7 +59,7 @@ async def create_invitation(
         secret_key=settings.INVITATIONS_SECRET_KEY.get_secret_value().encode(),
         base_url=settings.INVITATIONS_OSPARC_URL,
     )
-    invitation = _InvitationContentAndLink(
+    invitation = ApiInvitationContentAndLink(
         invitation_url=invitation_link,
         created=datetime.utcnow(),
         **invitation_inputs.dict(),
@@ -101,13 +72,13 @@ async def create_invitation(
 
 @router.post(
     "/invitations:extract",
-    response_model=_ApiInvitationContent,
+    response_model=ApiInvitationContent,
     response_model_by_alias=False,
 )
 async def extracts_invitation_from_code(
-    encrypted: _EncryptedInvitation,
+    encrypted: ApiEncryptedInvitation,
     settings: ApplicationSettings = Depends(get_settings),
-    _credentials: Optional[HTTPBasicCredentials] = Depends(get_validated_credentials),
+    _credentials: HTTPBasicCredentials | None = Depends(get_validated_credentials),
 ):
     """Decrypts the invitation code and returns its content"""
